+++ test-projects/login/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/cypress-project-login
+
 # @keystone-alpha/cypress-project-login
 
 ## 1.5.6